Call appy edit rules directly with options.
Refactor tests to run directly on editRules.apply.

Make removeTfVersionMentions and skipSectionHeadersEdit match any path, as we'll specify the index file when calling g.editRules.apply
